requirements.txt
customtkinter
google-generativeai
requests
